+++ messagix/mercury.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"mime/multipart"
 	"net/textproto"
+	"time"
 
 	"github.com/google/go-querystring/query"
 	"github.com/rs/zerolog"
@@ -33,7 +34,7 @@ func (c *Client) SendMercuryUploadRequest(ctx context.Context, threadID int64, m
 	urlQueries := c.NewHttpQuery()
 	queryValues, err := query.Values(urlQueries)
 	if err != nil {
-		return nil, fmt.Errorf("failed to convert HttpQuery into query.Values for mercury upload: %v", err)
+		return nil, fmt.Errorf("failed to convert HttpQuery into query.Values for mercury upload: %w", err)
 	}
 
 	payloadQuery := queryValues.Encode()
@@ -51,17 +52,25 @@ func (c *Client) SendMercuryUploadRequest(ctx context.Context, threadID int64, m
 	h.Set("sec-fetch-mode", "cors")
 	h.Set("sec-fetch-site", "same-origin") // header is required
 
-	_, respBody, err := c.MakeRequest(url, "POST", h, payload, types.NONE)
-	if err != nil {
-		return nil, fmt.Errorf("failed to send MercuryUploadRequest: %v", err)
+	var attempts int
+	for {
+		attempts += 1
+		_, respBody, err := c.MakeRequest(url, "POST", h, payload, types.NONE)
+		if err != nil {
+			// MakeRequest retries itself, so bail immediately if that fails
+			return nil, fmt.Errorf("failed to send MercuryUploadRequest: %w", err)
+		}
+		resp, err := c.parseMercuryResponse(ctx, respBody)
+		if err == nil {
+			return resp, nil
+		} else if attempts > MaxHTTPRetries {
+			return nil, err
+		}
+		c.Logger.Err(err).
+			Str("url", url).
+			Msg("Mercury response parsing failed, retrying")
+		time.Sleep(time.Duration(attempts) * 3 * time.Second)
 	}
-
-	resp, err := c.parseMercuryResponse(ctx, respBody)
-	if err != nil {
-		return nil, err
-	}
-
-	return resp, nil
 }
 
 var antiJSPrefix = []byte("for (;;);")
@@ -77,7 +86,7 @@ func (c *Client) parseMercuryResponse(ctx context.Context, respBody []byte) (*ty
 
 	var mercuryResponse *types.MercuryUploadResponse
 	if err := json.Unmarshal(jsonData, &mercuryResponse); err != nil {
-		return nil, fmt.Errorf("failed to parse mercury response: %v", err)
+		return nil, fmt.Errorf("failed to parse mercury response: %w", err)
 	} else if mercuryResponse.ErrorCode != 0 {
 		return nil, fmt.Errorf("error in mercury upload: %w", &mercuryResponse.ErrorResponse)
 	}
@@ -102,14 +111,14 @@ func (c *Client) parseMetadata(response *types.MercuryUploadResponse) error {
 		var realMetadata []types.FileMetadata
 		err := json.Unmarshal(response.Payload.Metadata, &realMetadata)
 		if err != nil {
-			return fmt.Errorf("failed to unmarshal image metadata in upload response: %v", err)
+			return fmt.Errorf("failed to unmarshal image metadata in upload response: %w", err)
 		}
 		response.Payload.RealMetadata = &realMetadata[0]
 	case '{':
 		var realMetadata map[string]types.FileMetadata
 		err := json.Unmarshal(response.Payload.Metadata, &realMetadata)
 		if err != nil {
-			return fmt.Errorf("failed to unmarshal video metadata in upload response: %v", err)
+			return fmt.Errorf("failed to unmarshal video metadata in upload response: %w", err)
 		}
 		realMetaEntry := realMetadata["0"]
 		response.Payload.RealMetadata = &realMetaEntry
@@ -127,24 +136,24 @@ func (c *Client) NewMercuryMediaPayload(media *MercuryUploadMedia) ([]byte, stri
 
 	err := writer.SetBoundary("----WebKitFormBoundary" + methods.RandStr(16))
 	if err != nil {
-		return nil, "", fmt.Errorf("messagix-mercury: Failed to set boundary (%v)", err)
+		return nil, "", fmt.Errorf("messagix-mercury: Failed to set boundary (%w)", err)
 	}
 
 	if media.IsVoiceClip {
 		err = writer.WriteField("voice_clip", "true")
 		if err != nil {
-			return nil, "", fmt.Errorf("messagix-mercury: Failed to write voice_clip field (%v)", err)
+			return nil, "", fmt.Errorf("messagix-mercury: Failed to write voice_clip field (%w)", err)
 		}
 
 		if media.WaveformData != nil {
 			waveformBytes, err := json.Marshal(media.WaveformData)
 			if err != nil {
-				return nil, "", fmt.Errorf("messagix-mercury: Failed to marshal waveform (%v)", err)
+				return nil, "", fmt.Errorf("messagix-mercury: Failed to marshal waveform (%w)", err)
 			}
 
 			err = writer.WriteField("voice_clip_waveform_data", string(waveformBytes))
 			if err != nil {
-				return nil, "", fmt.Errorf("messagix-mercury: Failed to write waveform field (%v)", err)
+				return nil, "", fmt.Errorf("messagix-mercury: Failed to write waveform field (%w)", err)
 			}
 		}
 	}
@@ -155,17 +164,17 @@ func (c *Client) NewMercuryMediaPayload(media *MercuryUploadMedia) ([]byte, stri
 
 	mediaPart, err := writer.CreatePart(partHeader)
 	if err != nil {
-		return nil, "", fmt.Errorf("messagix-mercury: Failed to create multipart writer (%v)", err)
+		return nil, "", fmt.Errorf("messagix-mercury: Failed to create multipart writer (%w)", err)
 	}
 
 	_, err = mediaPart.Write(media.MediaData)
 	if err != nil {
-		return nil, "", fmt.Errorf("messagix-mercury: Failed to write data to multipart section (%v)", err)
+		return nil, "", fmt.Errorf("messagix-mercury: Failed to write data to multipart section (%w)", err)
 	}
 
 	err = writer.Close()
 	if err != nil {
-		return nil, "", fmt.Errorf("messagix-mercury: Failed to close multipart writer (%v)", err)
+		return nil, "", fmt.Errorf("messagix-mercury: Failed to close multipart writer (%w)", err)
 	}
 
 	return mercuryPayload.Bytes(), writer.FormDataContentType(), nil